Create mactest.js
ASP Script HMAC generation example using JavaScript. Only run this script on the server. The code examples that are provided on the Adyen support department are offered as a support tool and cannot be guaranteed to be fit for every merchant specific situation. It is the merchant's own responsibility to implement the Adyen payment solution using these examples.